backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "argon2-cffi>=25.1.0",
    "dashscope>=1.24.7",
    "fastapi>=0.119.1",
    "langchain>=1.0.2",
    "langchain-chroma>=0.1.2",
    "langchain-community>=0.4",
    "langchain-milvus>=0.2.2",
    "langchain-openai>=1.0.1",
    "llama-index>=0.14.5",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
